runtimes

+++ code/datums/weakrefs.dm
@@ -18,7 +18,8 @@
 
 /datum/weakref/Destroy(force)
 	var/datum/target = resolve()
-	qdel(target)
+	if(target)
+		qdel(target)
 
 	if(!force)
 		return QDEL_HINT_LETMELIVE	//Let BYOND autoGC thiswhen nothing is using it anymore.
